Define Exposure Histories from an Exposure Profile

This function builds a matrix of exposure histories given an exposure profile, the time points at which each exposure history is evaluated, and a lag period.

exphist(exp, times, lag, fill=0)

Arguments

exp

an exposure profile defined at equally-spaced time units, from time 1 on.

times

either a numeric scalar or vector of integer numbers specifying the time points at which each exposure history is evaluated. By default, all the time points of exp. See Details.

lag

either an integer scalar or vector of length 2, defining the the maximum lag or the lag range, respectively. By default, the lag period from 0 to length(exp)-1.

fill

value used to fill the exposure history. See Details.

Details

This function is used to define matrices of exposure histories (backward in time) given an exposure profile (forward in time). Among other uses, this can be applied to define specific exposure histories for obtaining predictions in crosspred.

The exposure profile in exp is assumed to represent a series of exposure events defined forward in time, starting from time 1 and on. An exposure history is then evaluated backward in time for each point defined by times (rounded to integers) on the lag period defined by lag.

Negative numbers in exp represent time points before the start of the exposure profile, with 0 as the time immediately before, -1 as two times before, and so on. If the values in times are higher than the length of exp, or negative, or if the lag period extends backward before the beginning of the exposure profile, the exposure history is padded with values defined by fill.

Value

A numeric matrix of exposure histories, with named rows corresponding to the values in times and named columns corresponding to the lag period in lag.

References

Gasparrini A. Modeling exposure-lag-response associations with distributed lag non-linear models. Statistics in Medicine. 2014;33(5):881-899. [freely available here]

Author

Antonio Gasparrini <antonio.gasparrini@lshtm.ac.uk>

See also

Examples

### an exposure history evaluated at a single time
(exp <- sample(1:10))
#>  [1]  6  2 10  5  8  3  1  7  9  4
exphist(exp, 5, 3)
#>   lag0 lag1 lag2 lag3
#> 5    8    5   10    2
exphist(exp, 5, 12)
#>   lag0 lag1 lag2 lag3 lag4 lag5 lag6 lag7 lag8 lag9 lag10 lag11 lag12
#> 5    8    5   10    2    6    0    0    0    0    0     0     0     0
exphist(exp, 15, 3)
#>    lag0 lag1 lag2 lag3
#> 15    0    0    0    0

### use of argument lag
exphist(exp, 10, c(3,7))
#>    lag3 lag4 lag5 lag6 lag7
#> 10    1    3    8    5   10

### exposure histories evaluated at multiple times
exphist(exp, 3:5, 12)
#>   lag0 lag1 lag2 lag3 lag4 lag5 lag6 lag7 lag8 lag9 lag10 lag11 lag12
#> 3   10    2    6    0    0    0    0    0    0    0     0     0     0
#> 4    5   10    2    6    0    0    0    0    0    0     0     0     0
#> 5    8    5   10    2    6    0    0    0    0    0     0     0     0
exphist(exp, lag=12)
#>    lag0 lag1 lag2 lag3 lag4 lag5 lag6 lag7 lag8 lag9 lag10 lag11 lag12
#> 1     6    0    0    0    0    0    0    0    0    0     0     0     0
#> 2     2    6    0    0    0    0    0    0    0    0     0     0     0
#> 3    10    2    6    0    0    0    0    0    0    0     0     0     0
#> 4     5   10    2    6    0    0    0    0    0    0     0     0     0
#> 5     8    5   10    2    6    0    0    0    0    0     0     0     0
#> 6     3    8    5   10    2    6    0    0    0    0     0     0     0
#> 7     1    3    8    5   10    2    6    0    0    0     0     0     0
#> 8     7    1    3    8    5   10    2    6    0    0     0     0     0
#> 9     9    7    1    3    8    5   10    2    6    0     0     0     0
#> 10    4    9    7    1    3    8    5   10    2    6     0     0     0

### fill with NA's
exphist(exp, lag=12, fill=NA)
#>    lag0 lag1 lag2 lag3 lag4 lag5 lag6 lag7 lag8 lag9 lag10 lag11 lag12
#> 1     6   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#> 2     2    6   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#> 3    10    2    6   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#> 4     5   10    2    6   NA   NA   NA   NA   NA   NA    NA    NA    NA
#> 5     8    5   10    2    6   NA   NA   NA   NA   NA    NA    NA    NA
#> 6     3    8    5   10    2    6   NA   NA   NA   NA    NA    NA    NA
#> 7     1    3    8    5   10    2    6   NA   NA   NA    NA    NA    NA
#> 8     7    1    3    8    5   10    2    6   NA   NA    NA    NA    NA
#> 9     9    7    1    3    8    5   10    2    6   NA    NA    NA    NA
#> 10    4    9    7    1    3    8    5   10    2    6    NA    NA    NA

### see the vignette dlnmExtended for further examples