Nested Case-Control Study with a Time-Varying Exposure and a Cancer Outcome

The data set contains simulated data from an hypothetical nested case-control study on the association between a time-varying occupational exposure and a cancer outcome. The study includes 300 risk sets, each with a case and a control matched by age year. The data on the exposure is collected on 5-year age intervals between 15 and 65 years.

data(nested)

Format

A data frame with 600 observations on the following 14 variables.

id: subject ID.

case: indicator for case (1) or control (0).

age: age of each subject.

riskset: risk set id.

exp15: yearly exposure in the age period 15-19 year.

exp20: yearly exposure in the age period 20-24 year.

...

exp60: yearly exposure in the age period 60-64 year.

Details

The exposure history for each subject (series of yearly exposures) can be recovered by expanding the values given in exp15-exp60, and then selecting the values backward from the age of the subject for a given lag period.

Source

These nested case-control data were extracted from a simulated cohort with 300 cases of cancer and a time-varying exposure.

Author

Antonio Gasparrini <antonio.gasparrini@lshtm.ac.uk>

See also